+++ docs/conf.py
@@ -236,31 +236,33 @@ def update_yaml_block(file):
         lines = f.readlines()
     lines = [l.strip("\n") for l in lines]
 
-    s, e = None, None
+    s, e, obj = [], [], []
     for i, line in enumerate(lines):
         match = re.match(r"\.\. BEGIN YAML TARGET: ([\w.]+)\s*", line)
         if match:
-            obj = match.group(1)
-            s = i
+            obj.append(match.group(1))
+            s.append(i)
         elif line.rstrip() == ".. END YAML TARGET":
-            e = i
-    assert s is not None and e is not None, f"Unable to replace YAML targets in {file}"
-    assert s < e, f"Unable to replace YAML targets in {file}"
+            e.append(i)
+    assert len(s) > 0 and len(e) > 0, f"Unable to replace YAML targets in {file}"
+    assert all(si < ei for si, ei in zip(s, e)), f"Unable to replace YAML targets in {file}"
 
-    if s + 1 == e:
-        lines.insert(e, "")
-        e += 1
+    for si, ei, obji in list(zip(s, e, obj))[::-1]:
+        if si + 1 == ei:
+            lines.insert(ei, "")
+            ei += 1
 
-    module_path, member_name = obj.rsplit('.', 1)
-    member_data = getattr(import_module(module_path), member_name)
+        module_path, member_name = obji.rsplit('.', 1)
+        member_data = getattr(import_module(module_path), member_name)
 
-    defaults = convert_config_description_dict(member_data, include_required=True)
-    code = yaml.safe_dump(defaults, indent=2, sort_keys=False)
-    comments = extract_descriptions_as_comments(member_data)
-    code = add_comments_to_yaml(code, comments)
+        defaults = convert_config_description_dict(member_data, for_docs=True)
+        # breakpoint()
+        code = yaml.safe_dump(defaults, indent=2, sort_keys=False)
+        comments = extract_descriptions_as_comments(member_data)
+        code = add_comments_to_yaml(code, comments)
 
-    to_replace = "\n.. code-block:: yaml\n\n" + indent(code.rstrip(), "    ") + "\n"
-    lines[s+1:e] = to_replace.split("\n")
+        to_replace = "\n.. code-block:: yaml\n\n" + indent(code.rstrip(), "    ") + "\n"
+        lines[si+1:ei] = to_replace.split("\n")
 
     with open(file, "w") as f:
         f.write("\n".join(lines) + "\n")