+++ otter/execute/__init__.py
@@ -3,21 +3,22 @@ Execution and grading internals for Otter-Grader
 """
 
 import json
-import itertools
 import inspect
+import itertools
 import nbformat
 
 from IPython import get_ipython
 
 from .execute_log import execute_log
-from .execute_notebook import execute_notebook, filter_ignored_cells
-from .execute_script import execute_script
-# from .results import GradingResults
+from .execute_notebook import execute_notebook
+from .transforms import filter_ignored_cells, script_to_notebook
 
 from ..test_files import GradingResults, NotebookMetadataOKTestFile, OKTestFile
 from ..utils import id_generator
 
+
 NBFORMAT_VERSION = 4
+
 
 def check(nb_or_test_path, test_name=None, global_env=None):
     """
@@ -55,21 +56,24 @@ def check(nb_or_test_path, test_name=None, global_env=None):
 
     return test
 
-def grade_notebook(notebook_path, *, tests_glob=None, name=None, ignore_errors=True, script=False, 
-    cwd=None, test_dir=None, seed=None, log=None, variables=None, plugin_collection=None):
+
+def grade_notebook(submission_path, *, tests_glob=None, name=None, ignore_errors=True, script=False, 
+    cwd=None, test_dir=None, seed=None, seed_variable=None, log=None, variables=None, 
+    plugin_collection=None):
     """
     Grade an assignment file and return grade information
 
     Args:
-        notebook_path (``str``): path to a single notebook or Python script
+        submission_path (``str``): path to a single notebook or Python script
         tests_glob (``list`` of ``str``, optional): paths to test files to run
         name (``str``, optional): initial environment name
         ignore_errors (``bool``, optional): whether errors in execution should be ignored
-        script (``bool``, optional): whether the ``notebook_path`` is a Python script
+        script (``bool``, optional): whether the ``submission_path`` is a Python script
         cwd (``str``, optional): working directory of execution to be appended to ``sys.path`` in 
             grading environment
         test_dir (``str``, optional): path to directory of tests in grading environment
         seed (``int``, optional): random seed for intercell seeding
+        seed_variable (``str``, optional): a variable name to override with the seed
         log (``otter.check.logs.Log``, optional): log from which to grade questions
         variables (``dict``, optional): map of variable names -> type string to check type of deserialized
             object to prevent arbitrary code from being put into the environment; ignored if log is ``None``
@@ -79,19 +83,14 @@ def grade_notebook(notebook_path, *, tests_glob=None, name=None, ignore_errors=T
     Returns:
         ``otter.test_files.GradingResults``: the results of grading
     """
-    # # ensure this is not being executed inside a notebook
-    # assert get_ipython() is None, "Cannot execute inside Jupyter Notebook"
-
     if not script:
-        try:
-            with open(notebook_path) as f:
-                nb = nbformat.read(f, as_version=NBFORMAT_VERSION)
-        except UnicodeDecodeError:
-            with open(notebook_path, encoding='utf-8') as f:
-                nb = nbformat.read(f, as_version=NBFORMAT_VERSION)
+        nb = nbformat.read(submission_path, as_version=NBFORMAT_VERSION)
+
     else:
-        with open(notebook_path) as f:
+        with open(submission_path) as f:
             nb = f.read()
+
+        nb = script_to_notebook(nb)
 
     if plugin_collection is not None:
         nb = plugin_collection.before_execution(nb)
@@ -110,11 +109,14 @@ def grade_notebook(notebook_path, *, tests_glob=None, name=None, ignore_errors=T
         initial_env["__name__"] = name
 
     if log is not None:
-        global_env = execute_log(nb, log, secret, initial_env, ignore_errors=ignore_errors, cwd=cwd, test_dir=test_dir, variables=variables)
-    elif script:
-        global_env = execute_script(nb, secret, initial_env, ignore_errors=ignore_errors, cwd=cwd, test_dir=test_dir, seed=seed)
+        global_env = execute_log(
+            nb, log, secret, initial_env, ignore_errors=ignore_errors, cwd=cwd, test_dir=test_dir, 
+            variables=variables)
+
     else:
-        global_env = execute_notebook(nb, secret, initial_env, ignore_errors=ignore_errors, cwd=cwd, test_dir=test_dir, seed=seed)
+        global_env = execute_notebook(
+            nb, results_array, initial_env, ignore_errors=ignore_errors, cwd=cwd, test_dir=test_dir, 
+            seed=seed, seed_variable=seed_variable)
 
     if plugin_collection is not None:
         plugin_collection.run("after_execution", global_env)
@@ -132,10 +134,11 @@ def grade_notebook(notebook_path, *, tests_glob=None, name=None, ignore_errors=T
             for tested in tested_set:
                 if tested in t or t in tested:     # e.g. if 'tests/q1.py' is in /srv/repo/lab01/tests/q1.py
                     include = False
+
             if include:
                 extra_tests.append(OKTestFile.from_file(t))
                 extra_tests[-1].run(global_env)
-        # extra_results = [t.run(global_env, include_grade=False) for t in extra_tests]
+
         tests_run += extra_tests
 
     results = GradingResults(tests_run)