+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**Unreleased:**
+
+* Refactored Otter's execution internals
+* Added intercell seeding via seed variables per [#346](https://github.com/ucbds-infra/otter-grader/issues/346)
+
 **v3.0.4:**
 
 * Fixed bug in Otter Assign format v1 for closing exports when the last question is manually-graded